init: Allow init.real usage inside a recovery image
The AOSP method of adding init_sony causes init to be renamed in
both the boot and recovery ramdisks. This means that when trying
to use AOSP recovery, init_sony currently fails to boot as it does
not attempt to rename init.real once it has extracted the recovery
ramdisk from the FOTAKernel partition.

Move ths rename to catch both ramdisks, checking if init.real still
exists and therefore needs renaming.

+++ init/init_main.cpp
@@ -146,13 +146,13 @@ int main(int argc, char** __attribute__((unused)) argv)
         // Android boot
         write_string(BOOT_TXT, "ANDROID BOOT", true);
         init_board.introduce_android();
+    }
 
-        // Rename Android init on boot
-        if (argc < 2)
-        {
-            unlink("/init");
-            rename("/init.real", "/init");
-        }
+    // Rename init ready for boot
+    if (file_exists("/init.real") && argc < 2)
+    {
+        unlink("/init");
+        rename("/init.real", "/init");
     }
 
     // Finish init outputs
